# storage/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel API Docs",
        "version": "0.0.1",
        "description": "Backend API documentation for DevCanvas BaaS."
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/databases/{database_id}/collections/{collection_id}/delete": {
            "post": {
                "operationId": "collection.destroy",
                "tags": [
                    "Collection"
                ],
                "parameters": [
                    {
                        "name": "database_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Success"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Unable to delete collection"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "success": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "const": "Collection not found"
                                                },
                                                "data": {
                                                    "type": "null"
                                                },
                                                "errors": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "success",
                                                "message",
                                                "data",
                                                "errors"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "success": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "const": "Database not found"
                                                },
                                                "data": {
                                                    "type": "null"
                                                },
                                                "errors": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "success",
                                                "message",
                                                "data",
                                                "errors"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/databases/{database_id}/delete": {
            "post": {
                "operationId": "database.destroy",
                "tags": [
                    "Database"
                ],
                "parameters": [
                    {
                        "name": "database_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Database deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Unable to delete database"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Database not found"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/auth/user/logout": {
            "post": {
                "operationId": "user.destroy",
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Successfully logged out"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "User not authenticated"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Authorization: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}